moved services
moved services and updated dependencies
