+++ lion_master/src/app/layout/rdf-modelling/modelling-container/modelling-components/isa88/isa88.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Isa88ModelService, ISA88Insert, ISA88Variables } from '../../../rdf-models/isa88Model.service';
 import { Vdi3682ModelService } from '../../../rdf-models/vdi3682Model.service';
 import { Tables } from '../../../utils/tables';
-import { DownloadService } from '../../../rdf-models/services/download.service';
+import { DownloadService } from '../../../../../shared/services/backEnd/download.service';
 
 import { DataLoaderService } from '../../../../../shared/services/dataLoader.service';
 import { take } from 'rxjs/operators';
