+++ lion_master/src/app/layout/rdf-modelling/modelling-container/modelling-components/vdi3682/vdi3682.component.ts
@@ -3,7 +3,7 @@ import { FormBuilder, FormGroup } from '@angular/forms';
 import { Validators } from '@angular/forms';
 
 import { Vdi3682ModelService, VDI3682DATA, VDI3682VARIABLES, VDI3682INSERT } from '../../../rdf-models/vdi3682Model.service';
-import { PrefixesService } from '../../../rdf-models/services/prefixes.service';
+import { PrefixesService } from '../../../../../shared/services/prefixes.service';
 
 import { DataLoaderService } from '../../../../../shared/services/dataLoader.service';
 import { MessagesService } from '../../../../../shared/services/messages.service';
