+++ lion_master/src/app/layout/rdf-modelling/rdf-models/dinen61360Model.service.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@angular/core';
 
-import { PrefixesService } from './services/prefixes.service';
-import { QueriesService } from './services/backEnd/queries.service';
-import { GraphOperationsService } from './services/backEnd/graphOperations.service';
+import { PrefixesService } from '../../../shared/services/prefixes.service';
+import { QueriesService } from '../../../shared/services/backEnd/queries.service';
+import { GraphOperationsService } from '../../../shared/services/backEnd/graphOperations.service';
 import { DataLoaderService } from '../../../shared/services/dataLoader.service';
-import { DownloadService } from '../rdf-models/services/download.service';
+import { DownloadService } from '../../../shared/services/backEnd/download.service';
 import { take } from 'rxjs/operators';
 import { Observable } from 'rxjs';
