--- lion_master/src/app/layout/rdf-modelling/rdf-models/services/backEnd/configuration.service.spec.ts
@@ -1,16 +1,0 @@
-/* tslint:disable:no-unused-variable */
-
-import { TestBed, async, inject } from '@angular/core/testing';
-import { ConfigurationService } from './configuration.service';
-
-describe('Service: Configuration', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [ConfigurationService]
-    });
-  });
-
-  it('should ...', inject([ConfigurationService], (service: ConfigurationService) => {
-    expect(service).toBeTruthy();
-  }));
-});
