--- lion_master/src/app/layout/rdf-modelling/rdf-models/services/backEnd/graphOperations.service.spec.ts
@@ -1,16 +1,0 @@
-/* tslint:disable:no-unused-variable */
-
-import { TestBed, async, inject } from '@angular/core/testing';
-import { GraphOperationsService } from './graphOperations.service';
-
-describe('Service: GraphOperations', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [GraphOperationsService]
-    });
-  });
-
-  it('should ...', inject([GraphOperationsService], (service: GraphOperationsService) => {
-    expect(service).toBeTruthy();
-  }));
-});
