--- lion_master/src/app/layout/rdf-modelling/rdf-models/services/backEnd/queries.service.spec.ts
@@ -1,16 +1,0 @@
-/* tslint:disable:no-unused-variable */
-
-import { TestBed, async, inject } from '@angular/core/testing';
-import { QueriesService } from './queries.service';
-
-describe('Service: Queries', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [QueriesService]
-    });
-  });
-
-  it('should ...', inject([QueriesService], (service: QueriesService) => {
-    expect(service).toBeTruthy();
-  }));
-});
