--- lion_master/src/app/layout/rdf-modelling/rdf-models/services/download.service.spec.ts
@@ -1,16 +1,0 @@
-/* tslint:disable:no-unused-variable */
-
-import { TestBed, async, inject } from '@angular/core/testing';
-import { DownloadService } from './download.service';
-
-describe('Service: Download', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [DownloadService]
-    });
-  });
-
-  it('should ...', inject([DownloadService], (service: DownloadService) => {
-    expect(service).toBeTruthy();
-  }));
-});
