--- lion_master/src/app/layout/rdf-modelling/rdf-models/services/prefixes.service.spec.ts
@@ -1,16 +1,0 @@
-/* tslint:disable:no-unused-variable */
-
-import { TestBed, async, inject } from '@angular/core/testing';
-import { PrefixesService } from './prefixes.service';
-
-describe('Service: Prefixes', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [PrefixesService]
-    });
-  });
-
-  it('should ...', inject([PrefixesService], (service: PrefixesService) => {
-    expect(service).toBeTruthy();
-  }));
-});
