+++ lion_master/src/app/shared/services/backEnd/configuration.service.ts
