+++ lion_master/src/app/shared/services/backEnd/queries.service.ts
@@ -4,9 +4,9 @@ import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';
 import { take } from 'rxjs/operators';
 
 import { PrefixesService } from '../prefixes.service';
-import { Tables } from '../../../utils/tables';
-import { DataLoaderService } from "../../../../../shared/services/dataLoader.service";
-import { MessagesService } from "../../../../../shared/services/messages.service";
+import { Tables } from '../../../layout/rdf-modelling/utils/tables';
+import { DataLoaderService } from "../dataLoader.service";
+import { MessagesService } from "../messages.service";
 import { ConfigurationService } from './configuration.service';
